finished recipe match with database reference

+++ src/main/java/edu/brown/cs/cookups/Recipe.java
@@ -43,7 +43,7 @@ public class Recipe {
   public void setName(String n) {
   }
 
-  public Recipe scale(int partySize) {
+  public Recipe scale(double partySize) {
     Recipe toReturn = this;
     List<Ingredient> ingreds = ingredients;
     List<Ingredient> scaledIngredients = new ArrayList<>();
